pyproject.toml
[tool.poetry]
name = "final"
version = "0.1.0"
description = ""
authors = ["Merve YILDIRIM <merve.yildirim@etu.univ-tours.fr>"]

[tool.poetry.dependencies]
python = "^3.10"
rich = "^13.1.0"
typer = "^0.7.0"
pyserde = {extras = ["all"], version = "^0.9.7"}
numpy = "^1.24.1"
pandas = "^1.5.2"
scikit-learn = "^1.2.0"
selenium = "^4.7.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
